requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
passlib==1.7.4
httpx==0.28.1
PyJWT==2.9.0
pydantic-settings==2.3.4

# Pin bcrypt for passlib compatibility
bcrypt==4.0.1
bcrypt==4.0.1
